requirements.txt
docker==7.1.0
Flask==3.0.2
gunicorn==23.0.0
h5py==3.9.0
Jinja2==3.1.3               
json5==0.9.25              
jsonify==0.5  
matplotlib==3.7.5
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
pillow==10.3.0
requests==2.31.0
torch==2.2.1
tqdm==4.66.2
ultralytics==8.1.29 
